requirements.txt
Pillow==5.4.1
pytesseract==0.2.6